Added timer functionality and WPM
The timer starts only when you start typing. The WPM is calculated once you have correctly typed the quote. For now, I am just simply using the formula (no. of words / time in minutes) to calculate WPM. 
I plan to make it so that the WPM is calculated and shown every 3ish seconds rather than at the end only